Remove additional trigger code in electron.js
Commented out lines 14-20  and 272-274. These lines pertained to triggers and event codes. Given lines 102-186 were commented out, the pop-up for eeg connection would not appear, and the aforementioned code would fail, causing the package version of the task to be suspended as a background process in windows. Commentig out these lines solved the issue and the app will run as expected.